[feat](cache) Configure the image cache directory for coil
