+++ composeApp/src/webMain/kotlin/world/hachimi/app/ui/util/CoilConfig.web.kt
@@ -1,0 +1,7 @@
+package world.hachimi.app.ui.util
+
+import coil3.disk.DiskCache
+
+actual fun getPlatformDiskCache(): DiskCache? {
+    return null
+}
